requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.28.0
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.12.4
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
